Split the init_process file
