+++ crates/libcontainer/src/process/init/mod.rs
@@ -1,0 +1,7 @@
+mod context;
+pub mod error;
+pub mod process;
+
+pub use process::container_init_process;
+
+type Result<T> = std::result::Result<T, error::InitProcessError>;
