+++ crates/libcontainer/src/process/mod.rs
@@ -3,10 +3,10 @@
 
 pub mod args;
 pub mod channel;
-pub mod container_init_process;
 pub mod container_intermediate_process;
 pub mod container_main_process;
 mod fork;
+pub mod init;
 pub mod intel_rdt;
 mod message;
 #[cfg(feature = "libseccomp")]
